+++ backend-node/src/schemas/index.ts
@@ -1,0 +1,7 @@
+export * from './deviceSchemas';
+export * from './gatewaySchemas';
+export * from './bettercapSchemas';
+export * from './interceptorSchemas';
+export * from './shieldSchemas';
+export * from './gamingSchemas';
+export * from './authSchemas';